Histogram labels, axes
Just added some more labels to the histograms in the EDA, and adjusted the ylim so the graphs could be compared directly. 

The actual modelling part looks really good! Just wondering whether we need to think about if it makes sense to remove age from the model, based on it presumably being an important predictor to include - both for predictive and explanatory purposes??