Task 1 - interactions, knitted file
Played a bit more with the EDA section for readability.
Completed the interaction testing part and put forward a suggested final-final model with 2 interaction terms.
Updated the diagnostic plots/stats section with the final model (incl 2 interactions) and our subsetted dataset.

Knitted the file to ensure that was working -- have commented out some sections that are incomplete/possibly for deletion.

All of the commentary needs to be updated